+++ src/main/java/net/kaczmarzyk/spring/data/jpa/domain/NotLikeIgnoreCase.java
@@ -15,15 +15,11 @@
  */
 package net.kaczmarzyk.spring.data.jpa.domain;
 
-import net.kaczmarzyk.spring.data.jpa.utils.CaseConversionHelper;
-import net.kaczmarzyk.spring.data.jpa.utils.QueryContext;
-
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
-import jakarta.persistence.criteria.Expression;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
-import java.util.Locale;
+import net.kaczmarzyk.spring.data.jpa.utils.QueryContext;
 
 /**
  * Filters with {@code path not like %pattern%} where-clause and ignores pattern case
@@ -31,11 +27,9 @@ import java.util.Locale;
  * @author Kacper Leśniak (Tratif sp. z o.o.)
  *
  */
-public class NotLikeIgnoreCase<T> extends NotLike<T> implements IgnoreCaseStrategyAware, LocaleAware {
+public class NotLikeIgnoreCase<T> extends LikeIgnoreCase<T> {
 
     private static final long serialVersionUID = 1L;
-    private IgnoreCaseStrategy ignoreCaseStrategy;
-    private Locale locale;
 
     public NotLikeIgnoreCase(QueryContext queryContext, String path, String... args) {
         super(queryContext, path, args);
@@ -43,63 +37,19 @@ public class NotLikeIgnoreCase<T> extends NotLike<T> implements IgnoreCaseStrate
 
     @Override
     public Predicate toPredicate(Root<T> root, CriteriaQuery<?> query, CriteriaBuilder builder) {
-        var converted = CaseConversionHelper.applyCaseConversion(
-            builder,
-            this.<String>path(root),
-            pattern,
-            ignoreCaseStrategy,
-            locale
-        );
-        return builder.not(builder.like(converted.column(), converted.value()));
-    }
-    
-    @Override
-    public void setIgnoreCaseStrategy(IgnoreCaseStrategy ignoreCaseStrategy) {
-        this.ignoreCaseStrategy = ignoreCaseStrategy;
-    }
-    
-    @Override
-    @Deprecated
-    public void setLocale(Locale locale) {
-        this.locale = locale;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = super.hashCode();
-        result = prime * result + ((ignoreCaseStrategy == null) ? 0 : ignoreCaseStrategy.hashCode());
-        result = prime * result + ((locale == null) ? 0 : locale.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (!super.equals(obj)) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        NotLikeIgnoreCase<?> other = (NotLikeIgnoreCase<?>) obj;
-        if (ignoreCaseStrategy != other.ignoreCaseStrategy) {
-            return false;
-        }
-        if (locale == null) {
-            if (other.locale != null) {
-                return false;
-            }
-        } else if (!locale.equals(other.locale)) {
-            return false;
-        }
-        return true;
+        Predicate likeIgnoreCase = super.toPredicate(root, query, builder);
+        return builder.not(likeIgnoreCase);
     }
 
     @Override
     public String toString() {
-        return "NotLikeIgnoreCase [pattern=" + pattern + ", path=" + path + ", ignoreCaseStrategy=" + ignoreCaseStrategy + ", locale=" + locale + "]";
+        return "NotLikeIgnoreCase[" +
+                "argument='" + argument + '\'' +
+                ", escapeChar='" + escapeChar + '\'' +
+                ", pattern='" + pattern + '\'' +
+                ", path='" + path + '\'' +
+                ", ignoreCaseStrategy=" + ignoreCaseStrategy +
+                ", locale=" + locale +
+                ']';
     }
 }