+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/NotLikeIgnoreCaseTest.java
@@ -17,14 +17,14 @@ package net.kaczmarzyk.spring.data.jpa.domain;
 
 import net.kaczmarzyk.spring.data.jpa.Customer;
 import net.kaczmarzyk.spring.data.jpa.IntegrationTestBase;
+import net.kaczmarzyk.spring.data.jpa.utils.CharEscaper;
 import nl.jqno.equalsverifier.EqualsVerifier;
 import nl.jqno.equalsverifier.Warning;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
-import java.util.Locale;
+import java.util.Set;
 
 import static net.kaczmarzyk.spring.data.jpa.CustomerBuilder.customer;
 import static org.assertj.core.api.Assertions.assertThat;
@@ -56,6 +56,21 @@ public class NotLikeIgnoreCaseTest extends IntegrationTestBase {
         assertThat(result)
                 .hasSize(1)
                 .containsOnly(margeSimpson);
+    }
+
+    @Test
+    public void usesCharEscaper() {
+        customer("any", "Char%").build(em);
+        customer("any", "Char_").build(em);
+
+        NotLikeIgnoreCase<Customer> spec = new NotLikeIgnoreCase<>(queryCtx, "lastName", "CHAR%");
+        spec.applyCharEscaper(new CharEscaper('\\', Set.of('%', '_')));
+
+        List<Customer> result = customerRepo.findAll(spec);
+
+        assertThat(result)
+                .extracting(Customer::getLastName)
+                .containsOnly("Simpson", "Szyslak", "Char_");
     }
 
     @Test